\name{EESPCA-package}
\alias{EESPCA-package}
\docType{package}
\title{
Eigenvectors
}
\description{
Implementation of Eigenvectors from Eigenvalues Sparse Principal Component Analysis (EESPCA).
}
\details{
\tabular{ll}{
Package: \tab EESPCA\cr
Type: \tab Package\cr
Version: \tab 0.7.0 \cr
Date: \tab 2021\cr
License: \tab GPL-2 \cr
}
}
\author{
H. Robert Frost
}
\references{
    \itemize{
    \item Frost, H. R. (2022). Eigenvectors from Eigenvalues Sparse Principal Component Analysis (EESPCA). Journal of Computational and Graphical Statistics.
    }
}
\note{
This work was supported by the National Institutes of Health grants K01LM012426, R21CA253408, P20GM130454,and P30CA023108.
}
\keyword{ package }

